+++ macintosh/sceneDocument/templates/materialTemplate.mm
@@ -183,7 +183,7 @@ static BOOL mSkySphereWritten;
 			[ds addTab];
 			[ds copyTabAndText:@"material {\n"];
 			[ds addTab];
-			for (int x=beginLayer; x<=endLayer; x++)
+			for (NSInteger x=beginLayer; x<=endLayer; x++)
 				[MaterialTemplate addLayer:ds atIndex:x fromMap:cmap];
 			[ds removeTab];
 			if ( [[dict objectForKey:@"materialTransformationsOn"]intValue]==NSOnState)
@@ -203,7 +203,7 @@ static BOOL mSkySphereWritten;
 			[ds copyTabAndText:@"sphere {	0.0, 0.35}\n"];
 			[ds copyTabAndText:@"material {\n"]; [ds addTab];
 			[ds addTab];
-			for (int x=beginLayer; x<=endLayer; x++)
+			for (NSInteger x=beginLayer; x<=endLayer; x++)
 				[MaterialTemplate addLayer:ds atIndex:x fromMap:cmap];
 			if ( [[dict objectForKey:@"materialTransformationsOn"]intValue]==NSOnState)
 			{
@@ -223,7 +223,7 @@ static BOOL mSkySphereWritten;
 			[ds copyTabAndText:@"sphere {	<0.0, 0.0, 0.0>, 0.35}\n"];
 			[ds copyTabAndText:@"material {\n"];
 			[ds addTab];
-			for (int x=beginLayer; x<=endLayer; x++)
+			for (NSInteger x=beginLayer; x<=endLayer; x++)
 				[MaterialTemplate addLayer:ds atIndex:x fromMap:cmap];
 			if ( [[dict objectForKey:@"materialTransformationsOn"]intValue]==NSOnState)
 			{
@@ -241,7 +241,7 @@ static BOOL mSkySphereWritten;
 			[ds addTab];
 			[ds copyTabAndText:@"material {\n"]; 
 			[ds addTab];
-			for (int x=beginLayer; x<=endLayer; x++)
+			for (NSInteger x=beginLayer; x<=endLayer; x++)
 				[MaterialTemplate addLayer:ds atIndex:x fromMap:cmap];
 			if ( [[dict objectForKey:@"materialTransformationsOn"]intValue]==NSOnState)
 			{
@@ -259,7 +259,7 @@ static BOOL mSkySphereWritten;
 			[ds addTab];
 			[ds copyTabAndText:@"material {\n"];
 			[ds addTab];
-			for (int x=beginLayer; x<=endLayer; x++)
+			for (NSInteger x=beginLayer; x<=endLayer; x++)
 				[MaterialTemplate addLayer:ds atIndex:x fromMap:cmap];
 			if ( [[dict objectForKey:@"materialTransformationsOn"]intValue]==NSOnState)
 			{
@@ -646,7 +646,7 @@ static BOOL mSkySphereWritten;
 	[[self getWindow]makeFirstResponder: [self getWindow]];
 	id trimmedPrefs=nil;
 	NSString *PanelTitle=nil;
-	int currentTab=[materialMainTabView indexOfTabViewItem:[materialMainTabView selectedTabViewItem]];
+	NSInteger currentTab=[materialMainTabView indexOfTabViewItem:[materialMainTabView selectedTabViewItem]];
 	switch (currentTab)
 	{
 		case cMaterialPigmentTab:	
@@ -779,7 +779,7 @@ static BOOL mSkySphereWritten;
 //---------------------------------------------------------------------
 -(IBAction) materialTarget:(id)sender
 {
-	int theTag;
+	NSInteger theTag;
 	if ( sender==self)
 		theTag=cMaterialTransformationsOn;
 	else
@@ -838,7 +838,7 @@ static BOOL mSkySphereWritten;
 {
 	id 	prefs=nil;
 
-	int tag=[sender tag];
+	NSInteger tag=[sender tag];
 	
 	switch( tag)
 	{
