+++ macintosh/sceneDocument/templates/objectEditorTemplate.mm
@@ -489,17 +489,17 @@ enum {
 			NSInteger SelectedSegment=(rowIndex/4)+1;
 
 			if ( rowIndex == 0 )																//first point always start of segment
-				str=[NSString stringWithFormat:@"S %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"S %ld",(long)SelectedSegment];
 			else if ( rowIndex == [mMap count]-1)									//last point always end of a segment
-				str=[NSString stringWithFormat:@"E %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"E %ld",(long)SelectedSegment];
 			else if ( (((rowIndex+1)/2) & 1)	&& ((rowIndex+1) % 2==0))					// First controlpoint of a segment
-				str=[NSString stringWithFormat:@"Cs %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"Cs %ld",(long)SelectedSegment];
 			else if ( (rowIndex+1) % 4 == 0 )														// end of a segment
-				str=[NSString stringWithFormat:@"E %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"E %ld",(long)SelectedSegment];
 			else if ( rowIndex % 4 == 0)												//start of a segment
-				str=[NSString stringWithFormat:@"S %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"S %ld",(long)SelectedSegment];
 			else																						//must be last controlpoint of a segment
-				str=[NSString stringWithFormat:@"Ce %d",SelectedSegment];
+				str=[NSString stringWithFormat:@"Ce %ld",(long)SelectedSegment];
 		}
